book/requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
automium==0.2.6
automium-web==0.1.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cupshelpers==1.0
diff-match-patch==20121119
Django==1.11.16
django-filter==2.0.0
django-import-export==1.1.0
django-suit==0.2.26
djangorestframework==3.9.0
entrypoints==0.2.3.post1
et-xmlfile==1.0.1
Flask==1.0.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
iotop==0.6
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
keyring==12.2.1
keyrings.alt==3.0
legacy==0.1.6
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.9
parsel==1.5.0
pexpect==4.2.1
Pillow==5.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
PyDispatcher==2.0.5
pygobject==3.28.2
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pysmbc==1.0.15.6
python-apt==1.6.0
python-dateutil==2.7.5
pytz==2018.5
pyxdg==0.25
PyYAML==3.13
queuelib==1.5.0
reportlab==3.4.0
requests==2.18.4
Scrapy==1.5.1
SecretStorage==2.3.1
service-identity==17.0.0
six==1.11.0
tablib==0.12.1
Twisted==18.9.0
ufw==0.35
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.22
uWSGI==2.0.17.1
virtualenv==16.0.0
w3lib==1.19.0
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
